Viktorija Zakarauskaite
- **Comfortability [0 to 5]**
**3** - I found Project 3 to be the hardest one so far. I feel like there was not as much time to get comfortable with the pandas as we had the time with other 2 projects.
I am looking forward to going through all of the study materials one more time!

- **Completeness [0 to 5]**
**5** - I know that this is by far not a very perfect, extensive or advanced piece of work, but I am still happy with it as I started from 0 and fount many time contraints along the way.
Hooray for progress! Python is still super difficult though!

- **What was a win?**
To make the freaking visualizations work, I was really struggling with them for some reason!

- **What was a challenge?**
I found visuals to be difficult to apply!

- **Any other comments**
Just happy that I managed to deliver all the homework and projects no matter how hard it was for me and am very excited to continue learning more Python and further develop this skill!
Thank you to all of the team, I know that sometimes I was most likely not the easiest to teach as found this very very hard.
Thank you for your patience, support and understanding! Britt, Suresh and Vonn - all of you are great and thank you for everything! <3